requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
appdirs==1.4.4
asciichartpy==1.5.25
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
autobahn==21.3.1
black==21.5b0
blessings==1.7
cachetools==4.2.2
cbor==1.0.0
cbor2==5.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorful==0.5.4
commonmark==0.9.1
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.7
emoji==1.2.0
fastapi==0.64.0
filelock==3.0.12
flatbuffers==1.12
furl==2.1.2
google-api-core==1.26.3
google-auth==1.30.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.37.1
h11==0.12.0
hiredis==2.0.0
humanize==3.5.0
hyperlink==21.0.0
idna==2.10
inflect==5.3.0
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random
mkl-service==2.3.0
msgpack==1.0.2
multidict==5.1.0
munch==2.5.0
mypy==0.812
mypy-extensions==0.4.3
nltk==3.6.2
numba==0.53.1
numpy==1.20.2
nvidia-ml-py3==7.352.0
olefile==0.46
opencensus==0.7.12
opencensus-context==0.1.2
orderedmultidict==1.0.1
packaging==20.9
pathspec==0.8.1
Pillow
pooch==1.3.0
prometheus-client==0.10.1
protobuf==3.16.0
psutil==5.8.0
py-cpuinfo==8.0.0
py-spy==0.3.6
py-ubjson==0.16.1
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pyfiglet==0.8.post1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
resampy==0.2.2
rich==10.1.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
starlette==0.13.6
tensorboardX==2.2
text-unidecode==1.3
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.60.0
transformers==4.5.1
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==4.0.2
Unidecode==1.2.0
urllib3==1.26.4
uvicorn==0.13.4
wincertstore==0.2
xmltodict==0.12.0
yarl==1.6.3
